+++ config/config_test.go
@@ -476,6 +476,8 @@ func TestDefaultConfigWrite(t *testing.T) {
 ; httpd.retention = 6h0m0s
 ; httpd.timezones = 
 ; httpd.PIN.enabled = false
+; httpd.cards.default-start-date = 
+; httpd.cards.default-end-date = 
 
 # Wild Apricot
 ; wild-apricot.http.client-timeout = 10s
@@ -673,6 +675,8 @@ timeout = %[4]v
 httpd.retention = 5h30m0s
 ; httpd.timezones = 
 ; httpd.PIN.enabled = false
+; httpd.cards.default-start-date = 
+; httpd.cards.default-end-date = 
 
 # Wild Apricot
 ; wild-apricot.http.client-timeout = 10s