+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 2. Set default controller timezone from UTC to Local
 3. Added _httpd.security.admin.role_ to config.
 4. Updated to Go 1.23.
+5. Added default card start/end dates to HTTPD configuration.
 
 
 ## [0.8.8](https://github.com/uhppoted/uhppoted-lib/releases/tag/v0.8.8) - 2024-03-26